+++ src/Modal.php
@@ -115,7 +115,28 @@ class Modal extends Widget
      */
     public $toggleButton = false;
 
+    /**
+     * @var boolean whether to center the modal vertically
+     *
+     * When true the modal-dialog-centered class will be added to the modal-dialog
+     * @since 2.0.9
+     */
+    public $centerVertical = false;
+    /**
+     * @var boolean whether to make the modal body scrollable
+     *
+     * When true the modal-dialog-scrollable class will be added to the modal-dialog
+     * @since 2.0.9
+     */
+    public $scrollable = false;
 
+    /**
+     * @var array modal dialog options
+     * @see \yii\helpers\Html::renderTagAttributes() for details on how attributes are being rendered.
+     * @since 2.0.9
+     */
+    public $dialogOptions = [];
+    
     /**
      * Initializes the widget.
      */
@@ -124,10 +145,10 @@ class Modal extends Widget
         parent::init();
 
         $this->initOptions();
-
+       
         echo $this->renderToggleButton() . "\n";
         echo Html::beginTag('div', $this->options) . "\n";
-        echo Html::beginTag('div', ['class' => 'modal-dialog ' . $this->size]) . "\n";
+        echo Html::beginTag('div', $this->dialogOptions) . "\n";
         echo Html::beginTag('div', ['class' => 'modal-content']) . "\n";
         echo $this->renderHeader() . "\n";
         echo $this->renderBodyBegin() . "\n";
@@ -278,6 +299,18 @@ class Modal extends Widget
             if (!isset($this->toggleButton['data-target']) && !isset($this->toggleButton['href'])) {
                 $this->toggleButton['data-target'] = '#' . $this->options['id'];
             }
+        }
+        
+        $this->dialogOptions = array_merge([
+            'role'=>'document'
+        ], $this->dialogOptions);
+        Html::addCssClass($this->dialogOptions, 'modal-dialog');
+        Html::addCssClass($this->dialogOptions,  $this->size);
+        if ($this->centerVertical) {
+            Html::addCssClass($this->dialogOptions,  'modal-dialog-centered');
+        }
+        if ($this->scrollable) {
+            Html::addCssClass($this->dialogOptions,  'modal-dialog-scrollable');
         }
     }
 }